+++ src/templates/node-express-ts/babel.config.js
@@ -26,5 +26,5 @@ module.exports = {
       },
     ],
   ],
-  ignore: ['**/*.spec.ts'],
+  ignore: ['**/*.spec.ts', '**/*.test.ts'],
 };