+++ src/templates/node-express-ts/jest.config.ts
@@ -1,10 +1,19 @@
-import { pathsToModuleNameMapper } from 'ts-jest/utils';
-import { compilerOptions } from './tsconfig.json';
-
 export default {
-  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, {
-    prefix: '<rootDir>',
-  }),
+  clearMocks: true,
+  collectCoverage: true,
+  collectCoverageFrom: ['<rootDir>/src/modules/**/services/*.ts'],
+  coverageDirectory: 'coverage',
+  coverageReporters: ['text-summary', 'lcov'],
+  moduleNameMapper: {
+    '^@models/(.*)$': '<rootDir>/src/models/$1',
+    '^@views/(.*)$': '<rootDir>/src/views/$1',
+    '^@services/(.*)$': '<rootDir>/src/services/$1',
+    '^@controllers/(.*)$': '<rootDir>/src/controllers/$1',
+    '^@config/(.*)$': '<rootDir>/src/config/$1',
+    '^@repository/(.*)$': '<rootDir>/src/repository/$1',
+    '^@errors/(.*)$': '<rootDir>/src/errors/$1',
+  },
   preset: 'ts-jest',
   testEnvironment: 'node',
+  testMatch: ['**/*.spec.ts', '**/*.test.ts'],
 };