+++ src/templates/node-express-ts/src/errors/notFound.ts
@@ -1,11 +1,13 @@
 import { Request, Response, NextFunction } from 'express';
 
-export const notFoundHandler = (
+const notFoundHandler = (
   request: Request,
   response: Response,
-  next: NextFunction
+  _next: NextFunction
 ) => {
   const message = 'Resource not found';
 
   response.status(404).send(message);
 };
+
+export { notFoundHandler };